Move various transfer dashboard PT code utilities to their own module, lift up default PT value check to root component
